@@ -1,21 +1,18 @@
-package org.kushtrimhajrizi.rpalace.entity;
+package org.kushtrimhajrizi.rpalace.security.user;
 
-import org.kushtrimhajrizi.rpalace.security.PermissionType;
+import org.hibernate.annotations.GenericGenerator;
+import org.kushtrimhajrizi.rpalace.security.authority.Authority;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.core.userdetails.UserDetails;
 
-
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.JoinTable;
-import javax.persistence.ManyToMany;
+import javax.persistence.OneToMany;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.LinkedHashSet;
@@ -24,40 +21,35 @@ import java.util.Set;
 import java.util.stream.Collectors;
 
 @Entity
-public class DefaultUser implements UserDetails {
+public class User implements UserDetails {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    @Column(length = 64)
+    @GenericGenerator(name = "id_generator", strategy = "org.kushtrimhajrizi.rpalace.utils.IdGenerator")
+    @GeneratedValue(generator = "id_generator", strategy = GenerationType.SEQUENCE)
+    private String id;
     @Column(unique = true)
     private String email;
     @Column
     private String password;
-    @ManyToMany(cascade = {CascadeType.PERSIST, CascadeType.MERGE, CascadeType.REMOVE},
-            fetch = FetchType.EAGER)
-    @JoinTable(name = "user_permission",
-            joinColumns = @JoinColumn(name = "user_id"),
-            inverseJoinColumns = @JoinColumn(name = "permission")
-    )
-    private Set<Permission> permissions = new LinkedHashSet<>();
-    @Column
-    private String token;
+    @OneToMany(mappedBy = "user", fetch = FetchType.EAGER)
+    private Set<Authority> authorities = new LinkedHashSet<>();
     @Column
     private boolean enabled;
 
-    public static DefaultUser fromEmailAndPassword(String email, String password) {
-        DefaultUser user = new DefaultUser();
+    public static User fromEmailAndPassword(String email, String password) {
+        User user = new User();
         user.setEmail(email);
         user.setPassword(password);
         user.setEnabled(true);
         return user;
     }
 
-    public Integer getId() {
+    public String getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(String id) {
         this.id = id;
     }
 
@@ -83,30 +75,18 @@ public class DefaultUser implements UserDetails {
         return password;
     }
 
-    public Set<Permission> getPermissions() {
-        return permissions;
-    }
-
-    public void setPermissions(Set<Permission> permissions) {
-        this.permissions = permissions;
+    public void setAuthorities(Set<Authority> authorities) {
+        this.authorities = authorities;
     }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
-        if (permissions == null || permissions.isEmpty()) {
+        if (authorities == null || authorities.isEmpty()) {
             return Collections.emptyList();
         }
-        String[] permissionsArray = permissions.stream().map(Permission::getPermissionType)
-                .map(PermissionType::name).collect(Collectors.toList()).toArray(new String[]{});
+        String[] permissionsArray = authorities.stream().map(Authority::getAuthority)
+                .collect(Collectors.toList()).toArray(new String[]{});
         return AuthorityUtils.createAuthorityList(permissionsArray);
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
     }
 
     @Override
@@ -137,15 +117,14 @@ public class DefaultUser implements UserDetails {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        DefaultUser that = (DefaultUser) o;
+        User that = (User) o;
         return Objects.equals(id, that.id) &&
-                Objects.equals(email, that.email) &&
-                Objects.equals(token, that.token);
+                Objects.equals(email, that.email);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(id, email, token);
+        return Objects.hash(id, email);
     }
 
     @Override
@@ -154,8 +133,7 @@ public class DefaultUser implements UserDetails {
                 "id=" + id +
                 ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
-                ", permissions=" + permissions +
-                ", token='" + token + '\'' +
+                ", authorities=" + authorities +
                 ", enabled=" + enabled +
                 '}';
     }
