@@ -1,7 +1,6 @@
 package org.kushtrimhajrizi.rpalace.security.config;
 
-import org.kushtrimhajrizi.rpalace.entity.DefaultUser;
-import org.kushtrimhajrizi.rpalace.repository.UserRepository;
+import org.kushtrimhajrizi.rpalace.security.user.UserRepository;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -18,10 +17,7 @@ public class DefaultUserDetailsService implements UserDetailsService {
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        DefaultUser user = userRepository.findByEmail(email);
-        if (user == null) {
-            throw new UsernameNotFoundException("Email not found");
-        }
-        return user;
+        return userRepository.findByEmail(email)
+                .orElseThrow(() -> new UsernameNotFoundException("User not found"));
     }
 }
