@@ -1,13 +1,16 @@
-package org.kushtrimhajrizi.rpalace.service;
+package org.kushtrimhajrizi.rpalace.security.user;
 
-import org.kushtrimhajrizi.rpalace.dto.UserDTO;
 import org.kushtrimhajrizi.rpalace.exception.UserAlreadyExistsException;
-import org.kushtrimhajrizi.rpalace.exception.UserAlreadyHasToken;
-import org.kushtrimhajrizi.rpalace.exception.UserDoesNotExist;
+
+import java.util.Optional;
 
 public interface UserService {
 
     void save(UserDTO userDTO) throws UserAlreadyExistsException;
 
+    Optional<User> findByEmail(String email);
+
+/*
     void setToken(int id, String token) throws UserAlreadyHasToken, UserDoesNotExist;
+*/
 }
